# golang_backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Ufaelectrto market",
        "title": "Ufaelectrto API",
        "contact": {
            "name": "API Creator",
            "url": "https://t.me/amirich18",
            "email": "adamstradvers@gmail.com"
        },
        "version": "1.3"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/product_detail": {
            "get": {
                "description": "Get product detail with related products",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Product detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "prod_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.DefaultData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ProductInformation"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/products": {
            "get": {
                "description": "Get all products",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "All products",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product limit in response",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Response page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Category uuid",
                        "name": "cat_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Minimal products price",
                        "name": "price_min",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Maximum products price",
                        "name": "price_max",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Minimal products rating",
                        "name": "rating_min",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Maximum products rating",
                        "name": "rating_max",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Products uuid",
                        "name": "prod_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort settings: popular/lower_price/higher_price/news/discounts/default",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Is quantity are \u003e 0",
                        "name": "not_empty",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.CountResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "count": {
                                            "$ref": "#/definitions/dto.Count"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.Product"
                                            }
                                        },
                                        "sort_params": {
                                            "$ref": "#/definitions/dto.SortParams"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/products/review": {
            "post": {
                "description": "Create product review",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Create product review",
                "parameters": [
                    {
                        "description": "review info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.Review"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/products/reviews": {
            "get": {
                "description": "Get all product reviews",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get product reviews",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "prod_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Reviews limit in response",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Response page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.DefaultData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ProductStatistic"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.Category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "sub_categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SubCategory"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "dto.Characteristic": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "dto.Count": {
            "type": "object",
            "required": [
                "count"
            ],
            "properties": {
                "count": {
                    "type": "string"
                }
            }
        },
        "dto.CountResponse": {
            "type": "object",
            "properties": {
                "count": {},
                "data": {}
            }
        },
...
        "dto.DefaultData": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "dto.GetReview": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "rating": {
                    "type": "string"
                }
            }
        },
...
        "dto.Product": {
            "type": "object",
            "properties": {
                "base_unit": {
                    "type": "string"
                },
                "count": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                },
                "quantity": {
                    "type": "string"
                },
                "rating": {
                    "type": "string"
                },
                "review_count": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "vendor_code": {
                    "type": "string"
                }
            }
        },
        "dto.ProductFiles": {
            "type": "object",
            "properties": {
                "document": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "dto.ProductInformation": {
            "type": "object",
            "properties": {
                "base_unit": {
                    "type": "string"
                },
                "category": {
                    "$ref": "#/definitions/dto.Category"
                },
                "characteristics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Characteristic"
                    }
                },
                "count": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ProductFiles"
                    }
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                },
                "quantity": {
                    "type": "string"
                },
                "rating": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "vendor_code": {
                    "type": "string"
                }
            }
        },
        "dto.ProductStatistic": {
            "type": "object",
            "properties": {
                "rating": {
                    "type": "string"
                },
                "recommendation": {
                    "type": "string"
                },
                "review_count": {
                    "type": "string"
                },
                "reviews": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GetReview"
                    }
                },
                "sales_count": {
                    "type": "string"
                },
                "stars_statistic": {
                    "$ref": "#/definitions/dto.Stars"
                }
            }
        },
...
        "dto.Review": {
            "type": "object",
            "required": [
                "body",
                "product_id",
                "rating"
            ],
            "properties": {
                "body": {
                    "type": "string"
                },
                "image": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "product_id": {
                    "type": "string"
                },
                "rating": {
                    "type": "string"
                },
                "recommend": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.SortParams": {
            "type": "object",
            "properties": {
                "max_price": {
                    "type": "string"
                }
            }
        },
        "dto.Stars": {
            "type": "object",
            "properties": {
                "star_1": {
                    "type": "string"
                },
                "star_2": {
                    "type": "string"
                },
                "star_3": {
                    "type": "string"
                },
                "star_4": {
                    "type": "string"
                },
                "star_5": {
                    "type": "string"
                }
            }
        },
        "dto.SubCategory": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}